+++ orchestrator/Cargo.toml
@@ -12,7 +12,9 @@ serde = { version = "1.0.228", features = ["derive"] }
 tokio = { version = "1.52.1", features = ["full"] }
 anyhow = "1.0.102"
 serde_json.workspace = true
-util = { path = "../util"}
+async-trait = "0.1"
+tracing = "0.1.44"
 
 [dev-dependencies]
 tempfile = "3"
+mockall = "0.11"
