--- util/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "util"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-anyhow = "1.0.102"
-async-trait = "0.1"
-serde_json.workspace = true
-serde.workspace = true
-tokio = { version = "1.52.3", features = ["process"] }
-tracing = "0.1.44"
-
-[dev-dependencies]
-mockall = "0.11"
-tokio = { version = "1", features = ["macros", "rt-multi-thread"] }
